Relocate extra functions in erc4337_utils which are only needed for testing - move the utils package in tests.
